Top Tabs - Seperated viewcontroller from cells and moved buttons new file to DRY up the code
